requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pydantic==2.6.4
python-multipart==0.0.9
httpx==0.27.0
groq==0.5.0
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
groq